+++ backends/carp_webservices/lib/carp_services/protocol_service.dart
@@ -30,12 +30,16 @@ class CANSProtocolService extends CarpBaseService
   Future addVersion(StudyProtocol protocol, [String versionTag]) async =>
       StudyProtocol.fromJson(await _rpc(AddVersion(protocol, versionTag)));
 
+  /// Find all [StudyProtocol]'s owned by the owner with [ownerId].
+  /// In the CARP web service, the [ownerId] is the logged in user's [accountId].
+  ///
+  /// Returns the last version of each [StudyProtocol] owned by the requested owner,
+  /// or an empty list when none are found.
   @override
-  Future<List<StudyProtocol>> getAllFor(String ownerId) async {
-    Map<String, dynamic> responseJson = await _rpc(GetAllFor(ownerId));
-    List<dynamic> items = responseJson['items'];
-    return items.map((item) => StudyProtocol.fromJson(item)).toList();
-  }
+  Future<List<StudyProtocol>> getAllFor(String ownerId) async =>
+      (await _rpc(GetAllFor(ownerId)))['items']
+          .map((item) => StudyProtocol.fromJson(item))
+          .toList();
 
   @override
   Future<StudyProtocol> getBy(StudyProtocolId protocolId,