+++ carp_core/lib/carp_client/carp_core_client.dart
@@ -29,3 +29,4 @@ export 'carp_core_client.dart';
 part 'client_manager.dart';
 part 'study_runtime.dart';
 part 'device_manager.dart';
+part 'data_manager.dart';