+++ carp_core/lib/carp_deployment/domain/device_deployment.dart
@@ -39,6 +39,11 @@ class MasterDeviceDeployment {
   /// device deployment.
   DateTime lastUpdateDate;
 
+  // TODO - this is not part of carp_core Kotlin - make sure that no conflict arise.
+  //
+  /// Specifies where and how to upload the data collected from this deployment.
+  DataEndPoint dataEndPoint;
+
   MasterDeviceDeployment({
     this.deviceDescriptor,
     this.configuration,
@@ -47,6 +52,7 @@ class MasterDeviceDeployment {
     this.tasks,
     this.triggers,
     this.triggeredTasks,
+    this.dataEndPoint,
   }) {
     this.lastUpdateDate = DateTime.now();
   }