+++ carp_mobile_sensing/CHANGELOG.md
@@ -1,4 +1,9 @@
-## 0.21.3
+## 0.22.0
+* **BREAKING** - `DataEndPoint` is no longer specified as part of a `StudyProtocol`. 
+    * Specifying where and how to store or upload data is really not part of specifying a study protocol, but rather a deployment issues. Therefore, the `DataEndPoint` is now part of a `MasterDeviceDeployment`. 
+    * The `SmartphoneDeploymentService` will use the `FileDataEndpoint` per default if not specified in the deployment. Hence, data is saved as files on the phone.
+    * The `StudyDeploymentController.configure()` method now takes an optional `DataEndPoint` parameter to specify the data endpoint, if it shoudl be different to what is specified in the deployment.
+    * 
 * setting are now a singleton accessed using `Settings()`
 * better support for handling json functions